pcbnew: Fix strokeWidth truncation to integer mm in SVG polygons import The polygon conversion algorithm mistakenly stored the stroke width of filled polygons in an integer variable. As a result, all stroke widths were rounded down to the next smaller integer mm size. This commit changes the types in the relevant functions and variables from int to double.

+++ pcbnew/import_gfx/graphics_importer_buffer.h
@@ -188,7 +188,7 @@ public:
 
     std::vector<VECTOR2D>& Vertices() { return m_vertices; }
 
-    int GetWidth() const { return m_width; }
+    double GetWidth() const { return m_width; }
 
 private:
     std::vector<VECTOR2D> m_vertices;
